server.js: WhatsTheCode? serverside implementation
The idea is that there are rooms where people can share the latest "code"
for games like amongus, jackbox, etc. and not need to fiddle with looking at zoom chats,
text messages or other such mediums.

This commit implements a nodejs server listening on a configurable port that
listens for both http and websocket connections.

The http endpoints handle:
- the landing page
- creating a new room and redirecting to it

The websocket endpoint handles:
- establishing the connection and adding to a room based on the pathname of the request
- receiving incoming messages and broadcasting to the clients in the same room

TODO:
- websocket / room authorization needs to be cleaned up substantially. Its not secure at all
- broadcasting messages works for small rooms, but if this grows their needs to be a more effective approach
- the frontend pages all needs to be revamped and fluffed up
- rooms and clients are stored in memory as global variables. This should be persisted in a data store (db, cache, etc)